Create fault.c
This file is fault.c that detect page fault and determine it is going to be a user or kernel page fault so decide to do the page fault. I think if there is anything we want to customize it, we can use this one. But I think the best way might is to just add some debugging code or use its debug code. Or set a global value that store the address we want.